backend/requirements.txt
# Web Server & API
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# AI & Transcription
# Pinning to 1.0.3 for stability with the int8 compute type
faster-whisper==1.0.3
openai
requests
# torch is a dependency for some whisper internals, 
# but we only need the base package
torch --index-url https://download.pytorch.org/whl/cpu

# Media Processing
langdetect
ffmpeg-python

# System & Utilities
python-dotenv
httpx